Homepage redirect (#1104)
* initial redirect code

* fixed url update

* that's unnecessary

* removed featured code

* updated routes for revised urls

* bug fix
